+++ library/src/main/java/com/pacoworks/rxpaper2/RxPaperBook.java
@@ -19,7 +19,7 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.pacoworks.rxpaper;
+package com.pacoworks.rxpaper2;
 
 import android.content.Context;
 import android.util.Pair;
