+++ tests/src/main/AndroidManifest.xml
@@ -20,17 +20,17 @@
   -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.pacoworks.rxpaper.sample">
+    package="com.pacoworks.rxpaper2.sample">
 
     <application
-        android:name=".RxPaperSampleApplication"
+        android:name="com.pacoworks.rxpaper2.sample.RxPaperSampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.pacoworks.rxpaper.sample.MainActivity"
+            android:name="com.pacoworks.rxpaper2.sample.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar">
