+++ tests/src/main/res/layout/activity_main.xml
@@ -23,7 +23,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.pacoworks.rxpaper.sample.MainActivity">
+    tools:context="com.pacoworks.rxpaper2.sample.MainActivity">
 
     <TextView
         android:id="@+id/targetTextView"
