+++ tests/src/main/res/layout/content_main.xml
@@ -29,7 +29,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.pacoworks.rxpaper.sample.MainActivity"
+    tools:context="com.pacoworks.rxpaper2.sample.MainActivity"
     tools:showIn="@layout/activity_main">
 
     <TextView
