+++ tests/src/main/res/menu/menu_main.xml
@@ -22,7 +22,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.pacoworks.rxpaper.sample.MainActivity">
+    tools:context="com.pacoworks.rxpaper2.sample.MainActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
